apps/templates/services/python-flask/requirements.txt
ecs-logging==1.0.0
elasticsearch==8.0.0
flask==2.0.2
flask-cors==3.0.10
opentelemetry-api==1.9.1
opentelemetry-exporter-otlp==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-instrumentation-flask==0.28b1
opentelemetry-instrumentation-requests==0.28b1
psycopg2-binary==2.9.3
pymongo==3.11.3
redis==4.1.3
requests==2.27.1
